Generic autogam methods passed on to mgcv::gam methods

An autogam object contains a gam element that is simply an mgcv::gam object. So, it supports all mgcv::gam methods by, in most cases, simply passing the gam element on to their corresponding mgcv::gam methods. Only the following methods have special specifications for autogam (see their dedicated documentation files for details): print.autogam().

Usage

# S3 method for class 'autogam'
anova(x, ...)

# S3 method for class 'autogam'
coef(x, ...)

# S3 method for class 'autogam'
cooks.distance(model, ...)

# S3 method for class 'autogam'
formula(x, ...)

# S3 method for class 'autogam'
influence(model, ...)

# S3 method for class 'autogam'
logLik(x, ...)

# S3 method for class 'autogam'
model.matrix(x, ...)

# S3 method for class 'autogam'
predict(x, ...)

# S3 method for class 'autogam'
residuals(x, ...)

# S3 method for class 'autogam'
vcov(x, ...)

Arguments

x, model

An object of class autogam.

...

Additional arguments passed to other methods.

Value

Returns the return object of the corresponding mgcv::gam method.